Fix bfloat16→numpy crash: add .float() before .numpy() across all files
- Root cause: A100 autocast produces bfloat16 tensors, numpy rejects bfloat16
- Fixed torch.sigmoid(logits).cpu().numpy() → .float().cpu().numpy() in 11 locations
- Files: train.py, model.py, inference.py, retrain_trigger.py, benchmark_mps.py, NB03
- Also fixed total_mem → total_memory (PyTorch API) in train.py and NB03
- Tests: 20/20 passing

+++ src/benchmark_mps.py
@@ -107,7 +107,7 @@ def benchmark_mps():
             # Sample every 50th batch for metrics
             if batch_idx % 50 == 0:
                 with torch.no_grad():
-                    preds = torch.sigmoid(logits).cpu().numpy()
+                    preds = torch.sigmoid(logits).float().cpu().numpy()
                     sample_preds.extend(preds)
                     sample_labels.extend(labels.cpu().numpy())
